backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.12.2
pydantic-settings==2.4.0
pypdf==4.0.1
sap-ai-sdk-gen==5.9.0
langchain==0.3.25
langchain-community==0.3.25
